+++ packages/dma-library/src/operations/aave/v2/adjust-risk-down.ts
@@ -12,7 +12,7 @@ import {
 } from '@dma-library/types/operations'
 import BigNumber from 'bignumber.js'
 
-type AdjustRiskDownArgs = WithCollateralAndWithdrawal &
+export type AdjustRiskDownArgs = WithCollateralAndWithdrawal &
   WithDebt &
   WithOptionalDeposit &
   WithSwap &
@@ -41,14 +41,14 @@ export const adjustRiskDown: AaveV2AdjustDownOperation = async ({
   network,
 }) => {
   const setDaiApprovalOnLendingPool = actions.common.setApproval(network, {
-    amount: flashloan.amount,
+    amount: flashloan.token.amount,
     asset: addresses.DAI,
     delegate: addresses.lendingPool,
     sumAmounts: false,
   })
 
   const depositDaiInAAVE = actions.aave.v2.aaveDeposit(network, {
-    amount: flashloan.amount,
+    amount: flashloan.token.amount,
     asset: addresses.DAI,
     sumAmounts: false,
   })
@@ -92,7 +92,7 @@ export const adjustRiskDown: AaveV2AdjustDownOperation = async ({
 
   const withdrawDAIFromAAVE = actions.aave.v2.aaveWithdraw(network, {
     asset: addresses.DAI,
-    amount: flashloan.amount,
+    amount: flashloan.token.amount,
     to: addresses.operationExecutor,
   })
 
@@ -109,7 +109,7 @@ export const adjustRiskDown: AaveV2AdjustDownOperation = async ({
   const takeAFlashLoan = actions.common.takeAFlashLoan(network, {
     isDPMProxy: proxy.isDPMProxy,
     asset: addresses.DAI,
-    flashloanAmount: flashloan.amount,
+    flashloanAmount: flashloan.token.amount,
     isProxyFlashloan: true,
     provider: FlashloanProvider.DssFlash,
     calls: flashloanCalls,
