+++ packages/dma-library/src/protocols/aave/get-aave-protocol-data.ts
@@ -21,6 +21,7 @@ import { ethers, providers } from 'ethers'
 type InternalAaveProtocolData<AaveAddresses> = {
   collateralTokenAddress: string
   debtTokenAddress: string
+  flashloanTokenAddress: string
   proxy?: string
   addresses: AaveAddresses
   provider: providers.Provider
@@ -59,6 +60,7 @@ async function getAaveV2ProtocolData({
   provider,
   debtTokenAddress,
   collateralTokenAddress,
+  flashloanTokenAddress,
   proxy,
 }: InternalAaveProtocolData<AAVEStrategyAddresses> & { protocolVersion: AaveVersion.v2 }) {
   const priceOracle = new ethers.Contract(addresses.priceOracle, aaveV2PriceOracleABI, provider)
@@ -72,7 +74,7 @@ async function getAaveV2ProtocolData({
 
   const promises = [
     priceOracle
-      .getAssetPrice(addresses.DAI)
+      .getAssetPrice(flashloanTokenAddress)
       .then((amount: ethers.BigNumberish) => amountFromWei(new BigNumber(amount.toString()))),
     priceOracle
       .getAssetPrice(debtTokenAddress)
@@ -80,7 +82,7 @@ async function getAaveV2ProtocolData({
     priceOracle
       .getAssetPrice(collateralTokenAddress)
       .then((amount: ethers.BigNumberish) => amountFromWei(new BigNumber(amount.toString()))),
-    aaveProtocolDataProvider.getReserveConfigurationData(addresses.DAI),
+    aaveProtocolDataProvider.getReserveConfigurationData(flashloanTokenAddress),
     aaveProtocolDataProvider.getReserveConfigurationData(collateralTokenAddress),
   ]
 
@@ -92,7 +94,7 @@ async function getAaveV2ProtocolData({
   const results = await Promise.all(promises)
 
   return {
-    aaveFlashloanDaiPriceInEth: results[0] as BigNumber,
+    aaveFlashloanAssetPriceInEth: results[0] as BigNumber,
     aaveDebtTokenPriceInEth: results[1] as BigNumber,
     aaveCollateralTokenPriceInEth: results[2] as BigNumber,
     reserveDataForFlashloan: results[3],
@@ -109,6 +111,7 @@ async function getAaveV3ProtocolData({
   provider,
   debtTokenAddress,
   collateralTokenAddress,
+  flashloanTokenAddress,
   proxy,
 }: InternalAaveProtocolData<AAVEV3StrategyAddresses> & { protocolVersion: AaveVersion.v3 }) {
   const priceOracle = new ethers.Contract(
@@ -131,7 +134,7 @@ async function getAaveV3ProtocolData({
 
   const promises = [
     priceOracle
-      .getAssetPrice(addresses.DAI)
+      .getAssetPrice(flashloanTokenAddress)
       .then((amount: ethers.BigNumberish) => amountFromWei(new BigNumber(amount.toString()))),
     priceOracle
       .getAssetPrice(debtTokenAddress)
@@ -139,7 +142,7 @@ async function getAaveV3ProtocolData({
     priceOracle
       .getAssetPrice(collateralTokenAddress)
       .then((amount: ethers.BigNumberish) => amountFromWei(new BigNumber(amount.toString()))),
-    aaveProtocolDataProvider.getReserveConfigurationData(addresses.DAI),
+    aaveProtocolDataProvider.getReserveConfigurationData(flashloanTokenAddress),
     aaveProtocolDataProvider.getReserveConfigurationData(collateralTokenAddress),
   ]
 
@@ -174,7 +177,7 @@ async function getAaveV3ProtocolData({
   }
 
   return {
-    aaveFlashloanDaiPriceInEth: results[0] as BigNumber,
+    aaveFlashloanAssetPriceInEth: results[0] as BigNumber,
     aaveDebtTokenPriceInEth: results[1] as BigNumber,
     aaveCollateralTokenPriceInEth: results[2] as BigNumber,
     reserveDataForFlashloan: results[3],
