+++ packages/dma-library/src/strategies/aave/adjust/index.ts
@@ -1,2 +1,2 @@
-export { AaveAdjustArgs, AaveV2AdjustDependencies, AaveV3AdjustDependencies } from './adjust'
 export { adjust } from './adjust'
+export type { AaveAdjustArgs, AaveV2AdjustDependencies, AaveV3AdjustDependencies } from './types'
