+++ packages/dma-library/src/strategies/aave/close/close.ts
@@ -1,56 +1,16 @@
-import aavePriceOracleABI from '@abis/external/protocols/aave/v2/priceOracle.json'
-import aaveProtocolDataProviderABI from '@abis/external/protocols/aave/v2/protocolDataProvider.json'
-import aaveV3PriceOracleABI from '@abis/external/protocols/aave/v3/aaveOracle.json'
-import aaveV3ProtocolDataProviderABI from '@abis/external/protocols/aave/v3/aaveProtocolDataProvider.json'
-import { getForkedNetwork } from '@deploy-configurations/utils/network'
-import {
-  FEE_BASE,
-  FEE_ESTIMATE_INFLATOR,
-  ONE,
-  TEN,
-  TYPICAL_PRECISION,
-  ZERO,
-} from '@dma-common/constants'
-import { amountFromWei, amountToWei } from '@dma-common/utils/common'
+import { Network } from '@deploy-configurations/types/network'
+import { FEE_BASE, ONE, TEN, TYPICAL_PRECISION, ZERO } from '@dma-common/constants'
 import { calculateFee } from '@dma-common/utils/swap'
-import { operations } from '@dma-library/operations'
-import { AAVEStrategyAddresses } from '@dma-library/operations/aave/v2'
-import { AAVEV3StrategyAddresses } from '@dma-library/operations/aave/v3'
-import { AaveVersion } from '@dma-library/strategies/aave'
+import { buildOperation } from '@dma-library/strategies/aave/close/build-operation'
+import { generateTransition } from '@dma-library/strategies/aave/close/generate-transition'
+import { getValuesFromProtocol } from '@dma-library/strategies/aave/close/get-values-from-protocol'
 import { getAaveTokenAddresses } from '@dma-library/strategies/aave/get-aave-token-addresses'
-import {
-  IBasePositionTransitionArgs,
-  IOperation,
-  IPositionTransitionDependencies,
-  PositionTransition,
-  PositionType,
-  SwapData,
-  WithLockedCollateral,
-} from '@dma-library/types'
-import { AAVETokens } from '@dma-library/types/aave'
-import { resolveFlashloanProvider } from '@dma-library/utils/flashloan/resolve-provider'
+import { PositionTransition } from '@dma-library/types'
 import { acceptedFeeToken } from '@dma-library/utils/swap/accepted-fee-token'
 import { feeResolver } from '@dma-library/utils/swap/fee-resolver'
-import { FLASHLOAN_SAFETY_MARGIN, Position } from '@domain'
-import { Provider } from '@ethersproject/providers'
 import BigNumber from 'bignumber.js'
-import { ethers } from 'ethers'
 
-export type AaveCloseArgs = IBasePositionTransitionArgs<AAVETokens> & {
-  positionType: PositionType
-} & WithLockedCollateral & {
-    shouldCloseToCollateral?: boolean
-  }
-
-type WithVersioning = {
-  protocolVersion: AaveVersion
-}
-
-type AaveCloseArgsWithVersioning = AaveCloseArgs & WithVersioning
-
-export type AaveCloseDependencies =
-  | IPositionTransitionDependencies<AAVEStrategyAddresses>
-  | IPositionTransitionDependencies<AAVEV3StrategyAddresses>
+import { AaveCloseArgs, AaveCloseArgsWithVersioning, AaveCloseDependencies } from './types'
 
 export async function close(
   args: AaveCloseArgsWithVersioning,
@@ -62,8 +22,14 @@ export async function close(
 
   const { swapData, collectFeeFrom, preSwapFee } = await getSwapData(args, dependencies)
 
+  const flashloanToken =
+    dependencies.network === Network.MAINNET
+      ? dependencies.addresses.DAI
+      : dependencies.addresses.USDC
+
   const operation = await buildOperation(
     { ...swapData, collectFeeFrom, preSwapFee },
+    flashloanToken,
     args,
     dependencies,
   )
@@ -260,266 +226,4 @@ function calculateNeededCollateralToPaybackDebt(
       .div(colPrice),
   ).integerValue(BigNumber.ROUND_DOWN)
   return collateralAmountNeeded.times(TEN.pow(colPrecision - debtPrecision))
-}
-
-async function buildOperation(
-  swapData: SwapData & {
-    collectFeeFrom: 'sourceToken' | 'targetToken'
-    preSwapFee: BigNumber
-  },
-  args: AaveCloseArgsWithVersioning,
-  dependencies: AaveCloseDependencies,
-): Promise<IOperation> {
-  const { collateralTokenAddress, debtTokenAddress } = getAaveTokenAddresses(
-    { debtToken: args.debtToken, collateralToken: args.collateralToken },
-    dependencies.addresses,
-  )
-
-  const [aaveFlashloanDaiPriceInEth, aaveCollateralTokenPriceInEth, , reserveDataForFlashloan] =
-    await getValuesFromProtocol(
-      args.protocolVersion,
-      collateralTokenAddress,
-      debtTokenAddress,
-      dependencies,
-    )
-
-  /* Calculate Amount to flashloan */
-  const maxLoanToValueForFL = new BigNumber(reserveDataForFlashloan.ltv.toString()).div(FEE_BASE)
-  const ethPerDAI = new BigNumber(aaveFlashloanDaiPriceInEth.toString())
-  const ethPerCollateralToken = new BigNumber(aaveCollateralTokenPriceInEth.toString())
-  // EG STETH/ETH divided by ETH/DAI = STETH/ETH times by DAI/ETH = STETH/DAI
-  const oracleFLtoCollateralToken = ethPerCollateralToken.div(ethPerDAI)
-
-  const amountToFlashloanInWei = amountToWei(
-    amountFromWei(args.collateralAmountLockedInProtocolInWei, args.collateralToken.precision).times(
-      oracleFLtoCollateralToken,
-    ),
-    18,
-  )
-    .div(maxLoanToValueForFL.times(ONE.minus(FLASHLOAN_SAFETY_MARGIN)))
-    .integerValue(BigNumber.ROUND_DOWN)
-
-  const fee = feeResolver(args.collateralToken.symbol, args.debtToken.symbol)
-  const collateralAmountToBeSwapped = args.shouldCloseToCollateral
-    ? swapData.fromTokenAmount.plus(swapData.preSwapFee)
-    : args.collateralAmountLockedInProtocolInWei
-  const collectFeeFrom = swapData.collectFeeFrom
-  if (args.protocolVersion === AaveVersion.v2) {
-    const closeArgs = {
-      // In the close to collateral scenario we need to add the preSwapFee amount to the fromTokenAmount
-      // So, that when taking the fee from the source token we are sending the Swap contract
-      // the sum of the fee and the ultimately fromAmount that will be swapped
-      collateralAmountToBeSwapped,
-      flashloanAmount: amountToFlashloanInWei,
-      fee: fee.toNumber(),
-      swapData: swapData.exchangeCalldata,
-      receiveAtLeast: swapData.minToTokenAmount,
-      proxy: dependencies.proxy,
-      collectFeeFrom,
-      collateralTokenAddress,
-      collateralIsEth: args.collateralToken.symbol === 'ETH',
-      debtTokenAddress,
-      debtTokenIsEth: args.debtToken.symbol === 'ETH',
-      isDPMProxy: dependencies.isDPMProxy,
-      network: dependencies.network,
-    }
-    return await operations.aave.v2.close(
-      closeArgs,
-      dependencies.addresses as AAVEStrategyAddresses,
-    )
-  }
-  if (args.protocolVersion === AaveVersion.v3) {
-    const flashloanProvider = resolveFlashloanProvider(
-      await getForkedNetwork(dependencies.provider),
-    )
-
-    const closeArgs = {
-      collateral: {
-        address: collateralTokenAddress,
-        isEth: args.collateralToken.symbol === 'ETH',
-      },
-      debt: {
-        address: debtTokenAddress,
-        isEth: args.debtToken.symbol === 'ETH',
-      },
-      swap: {
-        fee: fee.toNumber(),
-        data: swapData.exchangeCalldata,
-        amount: collateralAmountToBeSwapped,
-        collectFeeFrom,
-        receiveAtLeast: swapData.minToTokenAmount,
-      },
-      flashloan: {
-        amount: amountToFlashloanInWei,
-        provider: flashloanProvider,
-      },
-      position: {
-        type: args.positionType,
-        collateral: { amount: collateralAmountToBeSwapped },
-      },
-      proxy: {
-        address: dependencies.proxy,
-        isDPMProxy: dependencies.isDPMProxy,
-        owner: dependencies.user,
-      },
-      addresses: dependencies.addresses as AAVEV3StrategyAddresses,
-      network: dependencies.network,
-    }
-
-    return await operations.aave.v3.close(closeArgs)
-  }
-
-  throw new Error('Unsupported AAVE version')
-}
-
-async function generateTransition(
-  swapData: SwapData,
-  collectFeeFrom: 'sourceToken' | 'targetToken',
-  preSwapFee: BigNumber,
-  operation: IOperation,
-  args: AaveCloseArgsWithVersioning,
-  dependencies: AaveCloseDependencies,
-) {
-  const currentPosition = dependencies.currentPosition
-  const { collateralTokenAddress, debtTokenAddress } = getAaveTokenAddresses(
-    { debtToken: args.debtToken, collateralToken: args.collateralToken },
-    dependencies.addresses,
-  )
-
-  const [, aaveCollateralTokenPriceInEth, aaveDebtTokenPriceInEth] = await getValuesFromProtocol(
-    args.protocolVersion,
-    collateralTokenAddress,
-    debtTokenAddress,
-    dependencies,
-  )
-  /*
-  Final position calculated using actual swap data and the latest market price
- */
-  const oracle = aaveCollateralTokenPriceInEth.div(aaveDebtTokenPriceInEth)
-  const finalPosition = new Position(
-    { amount: ZERO, symbol: currentPosition.debt.symbol },
-    { amount: ZERO, symbol: currentPosition.collateral.symbol },
-    oracle,
-    currentPosition.category,
-  )
-
-  const flags = { requiresFlashloan: true, isIncreasingRisk: false }
-
-  // We need to estimate the fee due when collecting from the target token
-  // We use the toTokenAmount given it's the most optimistic swap scenario
-  // Meaning it corresponds with the largest fee a user can expect to pay
-  // Thus, if the swap performs poorly the fee will be less than expected
-  const fromTokenAmountNormalised = amountFromWei(
-    swapData.fromTokenAmount,
-    args.collateralToken.precision,
-  )
-  const toTokenAmountNormalisedWithMaxSlippage = amountFromWei(
-    swapData.minToTokenAmount,
-    args.debtToken.precision,
-  )
-
-  const expectedMarketPriceWithSlippage = fromTokenAmountNormalised.div(
-    toTokenAmountNormalisedWithMaxSlippage,
-  )
-  const fee = feeResolver(args.collateralToken.symbol, args.debtToken.symbol)
-
-  const postSwapFee =
-    collectFeeFrom === 'targetToken' ? calculateFee(swapData.toTokenAmount, fee.toNumber()) : ZERO
-
-  return {
-    transaction: {
-      calls: operation.calls,
-      operationName: operation.operationName,
-    },
-    simulation: {
-      delta: {
-        debt: currentPosition.debt.amount.negated(),
-        collateral: currentPosition.collateral.amount.negated(),
-        flashloanAmount: ZERO,
-      },
-      flags: flags,
-      swap: {
-        ...swapData,
-        tokenFee: preSwapFee.plus(
-          postSwapFee.times(ONE.plus(FEE_ESTIMATE_INFLATOR)).integerValue(BigNumber.ROUND_DOWN),
-        ),
-        collectFeeFrom,
-        sourceToken: {
-          symbol: args.collateralToken.symbol,
-          precision: args.collateralToken.precision ?? TYPICAL_PRECISION,
-        },
-        targetToken: {
-          symbol: args.debtToken.symbol,
-          precision: args.debtToken.precision ?? TYPICAL_PRECISION,
-        },
-      },
-      position: finalPosition,
-      minConfigurableRiskRatio: finalPosition.minConfigurableRiskRatio(
-        expectedMarketPriceWithSlippage,
-      ),
-    },
-  }
-}
-
-async function getValuesFromProtocol(
-  protocolVersion: AaveVersion,
-  collateralTokenAddress: string,
-  debtTokenAddress: string,
-  dependencies: AaveCloseDependencies,
-) {
-  /* Grabs all the protocol level services we need to resolve values */
-  const { aavePriceOracle, aaveProtocolDataProvider } = getAAVEProtocolServices(
-    protocolVersion,
-    dependencies.provider,
-    dependencies.addresses,
-  )
-
-  // TODO: Add memoization
-  async function getAllAndMemoize() {
-    return Promise.all([
-      aavePriceOracle.getAssetPrice(dependencies.addresses.DAI),
-      aavePriceOracle.getAssetPrice(collateralTokenAddress),
-      aavePriceOracle.getAssetPrice(debtTokenAddress),
-      aaveProtocolDataProvider.getReserveConfigurationData(dependencies.addresses.DAI),
-    ])
-  }
-
-  return getAllAndMemoize()
-}
-
-function getAAVEProtocolServices(
-  protocolVersion: AaveVersion,
-  provider: Provider,
-  addresses: AAVEStrategyAddresses | AAVEV3StrategyAddresses,
-) {
-  switch (protocolVersion) {
-    case AaveVersion.v2:
-      return {
-        aavePriceOracle: new ethers.Contract(
-          (addresses as AAVEStrategyAddresses).priceOracle,
-          aavePriceOracleABI,
-          provider,
-        ),
-        aaveProtocolDataProvider: new ethers.Contract(
-          (addresses as AAVEStrategyAddresses).protocolDataProvider,
-          aaveProtocolDataProviderABI,
-          provider,
-        ),
-      }
-    case AaveVersion.v3:
-      return {
-        aavePriceOracle: new ethers.Contract(
-          (addresses as AAVEV3StrategyAddresses).aaveOracle,
-          aaveV3PriceOracleABI,
-          provider,
-        ),
-        aaveProtocolDataProvider: new ethers.Contract(
-          (addresses as AAVEV3StrategyAddresses).poolDataProvider,
-          aaveV3ProtocolDataProviderABI,
-          provider,
-        ),
-      }
-    default:
-      throw new Error('Unsupported AAVE Version')
-  }
 }
