+++ packages/dma-library/src/types/operations.ts
@@ -79,6 +79,10 @@ export type WithFlashloan = {
       amount: BigNumber
       address: Address
     }
+    /*
+     * @deprecated Please use `token` instead
+     */
+    amount: BigNumber
   }
 }
